+++ docs/changelog.md
@@ -5,6 +5,7 @@
 
 
 #### Module ryan_joiner
+- Added ``correlation_plot`` function (06/11/2023);
 - Added ``rj_test`` function (06/11/2023);
 - Added ``_p_value`` function (06/11/2023);
 - Added ``_statistic`` function (06/11/2023);